Track bcnt values to avoid declaring builds completed too early
When a package is rebuilt, it's sometimes possible for the previous
build status to not be updated for a few moments. During this time, if
the monitor jobs are started, they'll see a the final build stauts and
immediately complete...except that's not actually for the job we wanted
to monitor.

In order to fix this, track the bcnt (build count?) values that were
present pre-commit. Then, when the monitor gets a final build status,
it can check the history to ensure it's actually for the right build.
To make this work also involved refactoring moving srcmd5 retrival to
the upload stage, since it's needed to find the matching bcnt values.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>