Properly handle missing package build statuses
OBS seems to like to not update build metadata for a new revision for a
few moments after the upload completes.
141d7fa0734b291098db17f813e10197a29cd914 fixed this case for if a
previous build status was present, but this can also happen for a new
build (i.e. no previous build status was available). That should be
handled in the same way.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>