Reject old build statuses for non-rebuilds
141d7fa0734b291098db17f813e10197a29cd914 added support for tracking bcnt
values to avoid declaring a rebuilt package complete too soon due to the
old status sticking around. Although this fixed the issue in that case,
this can happen for *any* build submitted to a package that already has
some builds present, not just for explicit rebuilds. Thus, we now
perform the bcnt check for *all* builds being monitored.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>