Ensure we match the right srcmd5 when checking bcnt values
The previous entry in the bcnt saga,
3409f920f18229b4388fdb06a6621e41cf5c63da, starting tracking bcnt values
for all builds, not just rebuilds. In the case of all builds, however,
it's possible for the latest entry to *not* be a matching srcmd5;
because we only compared against the latest, bcnt values were matched
to the wrong build. Now, the latest build with an actually matching
srcmd5 is used, which was arguably the correct approach in the first
place.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>