Fix infinite wait for updated build status on failed builds
Because the build history only saves successful builds, failed builds
would never be added there, leading to the runner to wait forever for it
to appear. Instead, rely on the job history, which seems to record
successes *and* failures.

This also includes some slight improvements to the appearance of failed
builds.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>