Log errors as JSON, independently from GitLab
While thinking on #6 and the accidental / intentional use of
gitlab.output, it occurred to me that the way we show errors for GitLab
is not actually how we'd want them logged with JSON logging (which
currently dumps the entire error string, *including ANSI codes*, inside
the JSON objects). With this in mind, I realized we can get the
color-eyre handler's span traces and just convert it to JSON manually.

This would be all, if we could actually log JSON from tracing.
Unfortunately, any extra fields logged are converted to strings via
Debug and stored as such in the JSON fields. That means we end up with
our beautiful JSON errors stuck as strings and filled with dozens of
escape sequences, which also makes them entirely unreadable without
taking the JSON apart.

But! tracing has "unstable" (can change in a bugfix release, but isn't
that what Cargo.lock is for anyway?) functionality that integrates with
the valuable crate, such that any field value that's a Valuable will be
serialized as JSON. So, the eyre Report is converted to an intermediate
Valuable that gets sent to tracing, which dutifully serializes it as an
actual JSON object.

This does, however, still have an unfortunate missing piece, which is
serializing the fields in the span trace. tracing-error serializes
fields into a string to be retrieved when the span trace is collected,
which isn't the JSON format we would want. Now, this *can* be changed to
format JSON, but that's not very useful because:

- It means that span trace fields will be shown as JSON in the GitLab
  logs too, which isn't great.
- There is no straightforward way to go from a JSON string to a
  Valuable: valuable only includes ways to go serialize a Valuable via
  Serde, without any way of *deserializing*. This would only be possible
  by manually converting everything, which is increasing the
  already-high complexity.

In practice, the individual field values don't really have complex JSON
values, so leaving the default non-JSON formatting there is fine.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>