Set up JSON logging properly for fields
When using the JSON logging backend, tracing expects the fields to have
been formatted as a JSON string. However, the layer wasn't set up
properly for that, resulting in a variety of "field_error"s in the logs
(and panics on debug builds).

As it turns out, we were setting it up in an overcomplicated way
regardless, so use the simpler helper functions in tracing that set all
of this up for us.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>