Fix including excluded repositories due to broken status
There's a short period of time after a package is uploaded that OBS
might report it as `broken`, with the details set to `empty`. This means
that, when the build metadata is gathered, excluded architectures end up
being *included*, because the status is `broken` instead of `excluded`.

In order to work around this, wait for a short period of time if the
package's status is `broken` and details are `empty`, that way we can
actually check if the architecture is excluded.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>