Move filtering enabled repos to after the sources upload
When you branch an empty repository, the new repo ends up stuck in
'broken' state with status=empty. However, we now wait for the repo
status to *not* be broken / empty when getting the architecture list at
the start of an upload...which means that, in this case, the runner is
waiting forever for a non-empty repository that's never coming.

This is broken in the first place, though: architectures shouldn't be
examined until after the upload takes place and the repository is no
longer dirty, otherwise if the package changes the supported
architectures after an upload, the new list wouldn't be used. This moves
the architecture filtering to *after* the upload and waits for the dirty
status to ensure the list up-to-date. This fixes the infinite hangs,
because now when we go to check for broken / empty, the repository
itself should now have actual contents.

Fixes #12.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>