retry: be more robust against transient failures
Apertis' OBS server suffers from sometimes long(ish) periods of time
during which it issues 504. This causes the job to fail after 10 minutes
whereas it would succeed if we kept retrying.

As a temporary workardound, let's ignore the retry time limit for now,
so the job will keep trying until successful or the process is killed
(e.g. by cancelling the job).

Fixes #11

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>