Allow glbally setting the default monitor job timeout
PR #17 added support for job timeouts to the runner. This means that, if
your build jobs take a long time to complete, you must now override the
timeout via `--job-timeout`.

However, because the given option did not exist in previous releases,
deploying the updated version becomes more difficult: the new option is
not supported by the old version, but the moment you deploy the new
version, your builds might time out until the option is passed. To
circumvent this, this PR adds support for setting a global timeout for
all monitor jobs the runner generates, avoiding the need to urgently
make modifications to the pipelines.

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>