Move all fs access into a more generalized ArtifactDirectory trait
The current artifact handling & ArtifactDirectory have some issues:

- It only handles reading artifacts. This somewhat makes sense for the
  existing usage, but for future expansion purposes, there needs to be
  an abstraction for writing them, too. At the moment, artifacts are
  saved by random parts of the code creating temporary files and
  returning their handles all the way of to the job handler, which is
  absurdly inefficient for saving artifacts *to a known location*.
- The existing reading of artifacts is rather odd. There are 4 different
  methods that effectively are the product of:
  - `{returns None on missing, returns error on missing}`
  - `{returns AsyncFile, returns bytes}`
  which requires some silly code duplication and composes poorly.
- There's a lot of `file.try_clone()` expressions in contexts where it
  wants to separate the seek positions...except `.try_clone()` doesn't
  actually do that, and thus all that code is broken.

This completely overhauls all the artifact-related code to fix this:

- New methods are added to write a file via closure or to write some
  bytes directly.
- There's only one read method per *return type*, and errors related to
  it being missing are stuffed in a custom error, which has some utility
  types associated with it to make converting to an `Option` easy.
- A new trait is added to "reopen" files, which is used in the cases
  where the cursor positions need to actually be independent.

One odd quirk left behind is that reading and writing always deal in
`AsyncFile`s, which makes it impossible to easily write an
implementation that's e.g. memory-backend. I don't consider that to be a
major issue for now, because the alternative is writing a custom trait
for callers to be able to still `.try_clone()` (which isn't covered by
any existing async fs trait), and the only current use case is for the
tests, which...I'm fine with using redundant temporary files there for
now.