Refactor retry code on top of tokio-retry2 and macros
The `backoff` crate has been unmaintained for a while now, so this
switches to `tokio-retry2` as gitlab-runner-rs now uses:

https://github.com/collabora/gitlab-runner-rs/pull/90

Additionally, mixing async code with non-FnOnce (*especially* FnMut)
closures can be rather perilous, resulting a maze of extra locks and
clones. This is bad as-is in the runner, but `tokio-retry2` itself uses
an FnMut closure in its API, which makes everything worse. Instead, just
switch to a lightweight macro, which also makes it cleaner to pass in a
block.