Add a standalone CLI
This adds a new crate `obs-commander-cli` that behaves much like the
commands already available to the runner, but as a separate entity. As
replacement for child pipeline generation, the `generate-monitor`
command will create a JSON table containing the commands needed to run
monitoring and download binaries for a given repo/arch combination.