Remove the ScopedCell trick
The scoped cell trick was added to avoid lifetimes in the async callback
as they can be rather unergonomic and hard to figure out. However it
comes at the downside of moving a bunch of check to runtime.

The main trick here is to create a helper trait (Callback) , which
allows expressing bounds in ways that aren't possible in the `save_with`
generics. Having a generic implementation for the Callback trait allows
using closures naturally in other code.