artifacts: Add wrapper types for artifact reader and writers
Rather then exposing AsyncFile's directly to all users, create wrapper
types for r/w and read-only. When artifacts are created they use the
writer type which implements Async{Read,Write,Seek}, but cannot be
cloned. For storage and usage this is to be converted to a reader type
implement only Async{Read, Seek} such that the underlying file is
immutable.

Due to monitor already creating a handle within the save_with callback
this currently implements a `get_reader` call. Which means in principle
you can have both a writer and reader at the same time.. This is
something to be further refactored later.

The main benefit of this approach is that we're both avoiding unintended
write-access as well as making the underlying storage mostly opaque.